[meta] Move x86 registers generation to their own file;
